Undo extraction of shared DataSets code from importers
