+++ util/importers/librivox.py
@@ -3,6 +3,7 @@ import os
 import random
 import subprocess
 import tarfile
+import tensorflow as tf
 
 from glob import glob
 from itertools import cycle
@@ -12,15 +13,57 @@ from Queue import PriorityQueue
 from shutil import rmtree
 from tensorflow.contrib.learn.python.learn.datasets import base
 from tensorflow.python.platform import gfile
+from threading import Thread
 from util.audio import audiofile_to_input_vector
-from util.datasets import BaseDataSet, DataSets
-from util.text import text_to_char_array
+from util.gpu import get_available_gpus
+from util.text import text_to_char_array, ctc_label_dense_to_sparse
 
-class DataSet(BaseDataSet):
-    def __init__(self, *args, **kwargs):
-        super(DataSet, self).__init__(*args, **kwargs)
+class DataSets(object):
+    def __init__(self, train, dev, test):
+        self._dev = dev
+        self._test = test
+        self._train = train
+    
+    @property
+    def train(self):
+        return self._train
+    
+    @property
+    def dev(self):
+        return self._dev
+    
+    @property
+    def test(self):
+        return self._test
+
+class DataSet(object):
+    def __init__(self, session, txt_files, thread_count, batch_size, numcep, numcontext):
+        self._session = session
+        self._numcep = numcep
+        self._x = tf.placeholder(tf.float32, [None, numcep + (2 * numcep * numcontext)])
+        self._x_length = tf.placeholder(tf.int32, [])
+        self._y = tf.placeholder(tf.int32, [None,])
+        self._y_length = tf.placeholder(tf.int32, [])
+        self._example_queue = tf.PaddingFIFOQueue(shapes=[[None, numcep + (2 * numcep * numcontext)], [], [None,], []],
+                                                  dtypes=[tf.float32, tf.int32, tf.int32, tf.int32],
+                                                  capacity=2 * self._get_device_count() * batch_size)
+        self._enqueue_op = self._example_queue.enqueue([self._x, self._x_length, self._y, self._y_length])
+        self._txt_files = txt_files
+        self._batch_size = batch_size
+        self._numcontext = numcontext
+        self._thread_count = thread_count
         self._files_circular_list = self._create_files_circular_list()
         self._start_queue_threads()
+    
+    def _get_device_count(self):
+        available_gpus = get_available_gpus()
+        return max(len(available_gpus), 1)
+    
+    def _start_queue_threads(self):
+        batch_threads = [Thread(target=self._populate_batch_queue) for i in xrange(self._thread_count)]
+        for batch_thread in batch_threads:
+            batch_thread.daemon = True
+            batch_thread.start()
     
     def _create_files_circular_list(self):
         priorityQueue = PriorityQueue()
@@ -36,7 +79,7 @@ class DataSet(BaseDataSet):
     
     def _populate_batch_queue(self):
         for txt_file, wav_file in self._files_circular_list:
-            source = audiofile_to_input_vector(wav_file, self._num_mfcc_features, self._num_context)
+            source = audiofile_to_input_vector(wav_file, self._numcep, self._numcontext)
             source_len = len(next_source)
             with open(txt_file) as open_txt_file:
                 target = text_to_char_array(open_txt_file.read())
@@ -46,6 +89,17 @@ class DataSet(BaseDataSet):
                 self._x_length: source_len,
                 self._y: target,
                 self._y_length: target_len})
+
+    def next_batch(self):
+        source, source_lengths, target, target_lengths = self._example_queue.dequeue_many(self._batch_size)
+        sparse_labels = ctc_label_dense_to_sparse(target, target_lengths, self._batch_size)
+        return source, source_lengths, sparse_labels
+
+    @property
+    def total_batches(self):
+        # Note: If len(_txt_files) % _batch_size != 0, this re-uses initial _txt_files
+        return int(ceil(float(len(self._txt_files)) /float(self._batch_size)))
+
 
 def read_data_sets(session, data_dir, batch_size, numcep, numcontext, thread_count=8):
     # Check if we can convert FLAC with SoX before we start
